Audio-MW: For Long-Duration WMA(2 hours),Progress Bar is not in sync with playback after 1 hour of playback
BZ: 59271

Data-type conflict between the "presentationTime(ms)"(uint32_t) set by the AsfExtractor and
"mPositionTimeMediaUs(us)"(uint64_t) read from the input buffer in fillBuffer() of AudioPlayer
is causing this issue. In AsfExtractor.cpp,the "presentationTime"(ms) is type-casted to uint64_t
which resolves this issue.

Reviewed-on: http://android.intel.com:8080/71650
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Tested-by: buildbot <buildbot@intel.com>

+++ frameworks/asf_extractor/AsfExtractor.cpp
@@ -640,7 +640,7 @@ status_t AsfExtractor::readPacket() {
             buffer->set_range(0, payload->mediaObjectLength);
             // kKeyTime is in microsecond unit (usecs)
             // presentationTime is in mililsecond unit (ms)
-            buffer->meta_data()->setInt64(kKeyTime, payload->presentationTime * 1000);
+            buffer->meta_data()->setInt64(kKeyTime,(uint64_t) payload->presentationTime * 1000);
 
             if (payload->keyframe) {
                 buffer->meta_data()->setInt32(kKeyIsSyncFrame, 1);
@@ -657,7 +657,7 @@ status_t AsfExtractor::readPacket() {
                     Mutex::Autolock lockTrack(track->lock);
                     MediaBuffer* copy = NULL;
                     track->bufferPool->acquire_buffer(payload->payloadSize, &copy);
-                    copy->meta_data()->setInt64(kKeyTime, payload->presentationTime * 1000);
+                    copy->meta_data()->setInt64(kKeyTime,(uint64_t) payload->presentationTime * 1000);
                     memcpy(copy->data(), payload->payloadData, payload->payloadSize);
                     copy->set_range(0, payload->payloadSize);
                     track->bufferQueue.push(copy);
